third_party/blink/web_tests/accessibility/contenteditable-notifications.html

<!DOCTYPE html>
<script src="../resources/accessibility-helper.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id="textbox" contenteditable="true">
  First<p>Second</p>
</div>

<script>
'use strict';

async_test((t) => {
  // TODO(aboxhall): No longer necessary when updates are processed after layout.
  // Ensure entire a11y tree has already been seen.
  traverseAccessibilityTree(accessibilityController.rootElement);

  const textbox = document.getElementById('textbox');
  const axTextBox = accessibilityController.accessibleElementById('textbox');
  let valueChangedCount = 0;
  let selectedTextChangedCount = 0;
  const expectedIntents = [];

  axTextBox.setNotificationListener(t.step_func((notification, intents) => {
    if (notification == "ValueChanged") {
      ++valueChangedCount;
    } else if (notification == "SelectedTextChanged") {
      assert_array_equals(intents, [expectedIntents[selectedTextChangedCount]]);
      ++selectedTextChangedCount;
    }

    if (valueChangedCount >= 8 && selectedTextChangedCount >= 9) {
      t.done();
    }
  }));

  textbox.focus();
  // Initial setting of the selection.
  expectedIntents.push('AXEventIntent(setSelection,character,forward)');
  eventSender.keyDown("ArrowDown", []);
  expectedIntents.push('AXEventIntent(moveSelection,lineEnd,forward)');
  eventSender.keyDown("ArrowDown", []);
  expectedIntents.push('AXEventIntent(moveSelection,lineEnd,forward)');
  eventSender.keyDown("ArrowLeft", []);
  expectedIntents.push('AXEventIntent(moveSelection,character,backward)');
  eventSender.keyDown("ArrowLeft", []);
  expectedIntents.push('AXEventIntent(moveSelection,character,backward)');
  eventSender.keyDown("w", []);
  // TODO(nektar): Support an intent for the typing command.
  expectedIntents.push('AXEventIntent(setSelection,character,forward)');
  eventSender.keyDown("x", []);
  expectedIntents.push('AXEventIntent(setSelection,character,forward)');
  eventSender.keyDown("y", []);
  expectedIntents.push('AXEventIntent(setSelection,character,forward)');
  eventSender.keyDown("z", []);
  expectedIntents.push('AXEventIntent(setSelection,character,forward)');
}, 'This test ensures that moving the cursor in a contentEditable sends a selected text change notification, and typing in a contentEditable sends both a value changed and selected text changed notification.');

</script>
